pyproject.toml
[project]
name = "langgraph_intro"
version = "0.1.0"
description = ""
authors = [
    {name = "Divyanshu9382",email = "divyanshu9382@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=1.0.7,<2.0.0)",
    "langchain-core (>=1.0.5,<2.0.0)",
    "langchain-openai (>=1.0.3,<2.0.0)",
    "langchain-tavily (>=0.2.13,<0.3.0)",
    "langgraph (>=1.0.3,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-google-vertexai (>=3.0.3,<4.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
